Chore: fix file structure
