--- src/pages/Project.test.jsx
@@ -1,8 +1,0 @@
-import renderer from 'react-test-renderer';
-
-import Project from './Project';
-
-test('render the project page', () => {
-  const tree = renderer.create(<Project />).toJSON();
-  expect(tree).toMatchSnapshot();
-});
