+++ src/pages/Project/Project.test.jsx
@@ -1,0 +1,8 @@
+import renderer from 'react-test-renderer';
+
+import Project from './Project';
+
+test('render the project page', () => {
+  const tree = renderer.create(<Project />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
